+++ x/ecocredit/base/keeper/query_credit_class_fees.go
@@ -3,25 +3,31 @@ package keeper
 import (
 	"context"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	regentypes "github.com/regen-network/regen-ledger/types"
 	types "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
 )
 
-// CreditClassFees queries credit class creation fees.
-func (k Keeper) CreditClassFees(ctx context.Context, request *types.QueryCreditClassFeesRequest) (*types.QueryCreditClassFeesResponse, error) {
-	result, err := k.stateStore.ClassFeesTable().Get(ctx)
+// ClassFee queries credit class creation fees.
+func (k Keeper) ClassFee(ctx context.Context, request *types.QueryClassFeeRequest) (*types.QueryClassFeeResponse, error) {
+	classFee, err := k.stateStore.ClassFeeTable().Get(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	classFee, ok := regentypes.ProtoCoinsToCoins(result.Fees)
-	if !ok {
-		return nil, sdkerrors.ErrInvalidType.Wrap("credit class fee")
+	var fee sdk.Coin
+
+	if classFee.Fee != nil {
+		var ok bool
+		fee, ok = regentypes.ProtoCoinToCoin(classFee.Fee)
+		if !ok {
+			return nil, sdkerrors.ErrInvalidType.Wrapf("failed to parse class fee")
+		}
 	}
 
-	return &types.QueryCreditClassFeesResponse{
-		Fees: classFee,
+	return &types.QueryClassFeeResponse{
+		Fee: &fee,
 	}, nil
 }