+++ x/ecocredit/genesis/genesis.go
@@ -514,17 +514,17 @@ func MergeAllowedDenomsIntoTarget(messages []markettypes.AllowedDenom, target or
 	return w.Close()
 }
 
-// MergeCreditClassFeesIntoTarget merges params message into the ormjson.WriteTarget.
-func MergeCreditClassFeesIntoTarget(
+// MergeClassFeeIntoTarget merges params message into the ormjson.WriteTarget.
+func MergeClassFeeIntoTarget(
 	cdc codec.JSONCodec,
-	creditClassFees basetypes.ClassFees,
+	classFee basetypes.ClassFee,
 	target ormjson.WriteTarget) error {
-	w, err := target.OpenWriter(protoreflect.FullName(gogoproto.MessageName(&creditClassFees)))
+	w, err := target.OpenWriter(protoreflect.FullName(gogoproto.MessageName(&classFee)))
 	if err != nil {
 		return err
 	}
 
-	bz, err := cdc.MarshalJSON(&creditClassFees)
+	bz, err := cdc.MarshalJSON(&classFee)
 	if err != nil {
 		return err
 	}